Lifted up some assumptions that node IDs are always 64 bits

+++ benchmark/benchmarks/benchmark.h
@@ -569,7 +569,7 @@ void print_results_part(FILE* f, Graph* graph, int* a, int max=50) {
 	node_t m = std::min<node_t>(max, graph->max_nodes());
 
 	for (node_t n = 0; n < m; n++) {
-		if (n % 10 == 0) fprintf(f, "%7ld:", n);
+		if (n % 10 == 0) fprintf(f, "%7ld:", (long) n);
 		fprintf(f, " %7d", a[n]);
 		if (n % 10 == 9 || n + 1 == m) fprintf(f, "\n");
 	}
@@ -592,7 +592,7 @@ void print_results_part(FILE* f, Graph* graph, long* a, int max=50) {
 	node_t m = std::min<node_t>(max, graph->max_nodes());
 
 	for (node_t n = 0; n < m; n++) {
-		if (n % 10 == 0) fprintf(f, "%7ld:", n);
+		if (n % 10 == 0) fprintf(f, "%7ld:", (long) n);
 		fprintf(f, " %7ld", a[n]);
 		if (n % 10 == 9 || n + 1 == m) fprintf(f, "\n");
 	}
@@ -615,7 +615,7 @@ void print_results_part(FILE* f, Graph* graph, float* a, int max=50) {
 	node_t m = std::min<node_t>(max, graph->max_nodes());
 
 	for (node_t n = 0; n < m; n++) {
-		if (n % 10 == 0) fprintf(f, "%7ld:", n);
+		if (n % 10 == 0) fprintf(f, "%7ld:", (long) n);
 		fprintf(f, " %0.10f", a[n]);
 		if (n % 10 == 9 || n + 1 == m) fprintf(f, "\n");
 	}
@@ -638,7 +638,7 @@ void print_results_part(FILE* f, Graph* graph, double* a, int max=50) {
 	node_t m = std::min<node_t>(max, graph->max_nodes());
 
 	for (node_t n = 0; n < m; n++) {
-		if (n % 10 == 0) fprintf(f, "%7ld:", n);
+		if (n % 10 == 0) fprintf(f, "%7ld:", (long) n);
 		fprintf(f, " %0.10lf", a[n]);
 		if (n % 10 == 9 || n + 1 == m) fprintf(f, "\n");
 	}
